layout, selection and time and cs coloring
these update the layout of the page into a top div for buttons and then
three columns below, left for variables, middle for modeling, and right
for graphics.  users can select which variables they want to use in
their model and click layout.  also enabled are coloring by time and
cross section.  in the code, an attribute called nodeCol has been added
to the nodes, and eventually when passed to r the nodeCol attribute (or
field or whatever) of the node is how we can identify the type of
variable.