add/del nodes, graphics expansion, allNodes array, other program structuring
nodes are added and deleted by clicking on the variable name in the
left panel.  after piping in a figure to the right panel, mousing over
the figure will increase its size.  the previous layout features have
been removed, and now the layout is automatic but variable addition and
deletion is dynamic.  allNodes array has been established, and this
contains objects corresponding to each node.  allNodes is assigned when
the data are read in, and the barplots are also calculated at this
time. minor changes to the css file to allow the right div to resize
over main div (remove float of right div and add absolute positioning).