dvn link, tabs, and zelig model object
the ddi is now accessed through a link to the dvn build.  in the
future, this link will be passed to sola fide when the button is
clicked in dvn.

 a tab for estimators and variables has been added, eventually this is
to include all zelig estimators.  if the variable list is too long for
the leftpanel div, a scrollbar is used to scroll through them.

a zelig model object is defined when the "estimate" button is clicked.
this includes the dataset name, variables, edges, time and cross
section variable if selected, and the selected zelig estimator.