Logic for switching between local and dvn data
Added cleaner logic for switching between sample data supplied with the app (files stored 
in the "data" directory). If no file id is supplied, the app will load the DDI metadata from one of 
the local files. If it is supplied (as a URL argument, for ex., gui.html?dfId=17), the app 
will get the DDI from the dataverse specified. 
This way the app should continue working for the developers as is; with only minimal edits required to install it piggy-backed to a dataverse).