changed data file parameter in rookzelig
Changed the way rookzelig code receives the parameters for obtaining tab-delimited data. 
Instead of receiving the hostname and fileid, then cooking a dataverse file access url, it now 
receives a complete url (zdataurl). This way we don't have to assume that we are dealing
with a dataverse; the url could be pointing to any location.