modified rookzelig.R to handle https in data URLs
We discovered that R method read.delim() does not work with https URLs. 
So I worked around it with this quick hack - I first download the tab-delimited file and save it locally, using download.file() method; 
then read the local file with read.delim. 
If anyone knows of a prettier way of doing this - please change it accordingly...