Update package_versions.txt
Added all the packages that my R setup script tries to install, to the dependencies list that Vito started. 
(the R setup script in question is the one supplied with the Dataverse and mentioned in the installation instructions). 
TODO: check if all of these are still needed;  and if we need specific versions of the ones that are needed. 

For now my script is going to assume that if no version is specified, simply pulling the latest from CRAN should be ok.