Update install.pl
Changed the default values, to reflect the way Glassfish and Apache are configured out of the box. If this configuration is used, rApache will be running at http://hostname:80 and the Dataverse at http://hostname:8080. Added a section to the installation guide that discusses the port configuration, should they want to change these values.